requirements.txt
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.24.0
scipy>=1.11.0
wandb>=0.16.0
scikit-learn>=1.3.0
PyYAML>=6.0
pytest>=7.4.0
pytest-timeout>=2.2.0
tqdm>=4.66.0
einops>=0.7.0
